Move startOfDay and endOfDay to the shared date utilities
Both already existed inside SearchExpensesFilters, exported and tested, but
living in a feature folder meant nobody found them: two more copies were added
in DateQuickSelector and AddExpenseForm, plus one in each of their test files.

The test copies were the real problem. Both tests built their expected dates
with their own reimplementation of the logic under test, so a bug in
startOfDay would have been mirrored in the expectation and the tests would
still have passed.
