+++ src/components/Expenses/SearchExpensesFilters/utils.test.js
@@ -1,4 +1,4 @@
-import { parseAmount, startOfDay, endOfDay, buildSearchVariables, isValidAmountInput, buildFiltersSummary } from './utils'
+import { parseAmount, buildSearchVariables, isValidAmountInput, buildFiltersSummary } from './utils'
 
 const categories = [
 	{
@@ -48,38 +48,6 @@ describe('parseAmount', () => {
 
 	it('should return undefined for a value that is not a number', () => {
 		expect(parseAmount('abc')).toBeUndefined()
-	})
-})
-
-describe('startOfDay and endOfDay', () => {
-	it('should move a date to the very beginning of its day', () => {
-		const result = startOfDay(new Date(2026, 1, 1, 17, 42, 13, 500))
-
-		expect(result.getFullYear()).toBe(2026)
-		expect(result.getMonth()).toBe(1)
-		expect(result.getDate()).toBe(1)
-		expect(result.getHours()).toBe(0)
-		expect(result.getMinutes()).toBe(0)
-		expect(result.getSeconds()).toBe(0)
-		expect(result.getMilliseconds()).toBe(0)
-	})
-
-	it('should move a date to the very end of its day', () => {
-		const result = endOfDay(new Date(2026, 3, 30, 8, 5, 0, 0))
-
-		expect(result.getDate()).toBe(30)
-		expect(result.getHours()).toBe(23)
-		expect(result.getMinutes()).toBe(59)
-		expect(result.getSeconds()).toBe(59)
-		expect(result.getMilliseconds()).toBe(999)
-	})
-
-	it('should not mutate the date received', () => {
-		const original = new Date(2026, 1, 1, 17, 42, 13, 500)
-
-		startOfDay(original)
-
-		expect(original.getHours()).toBe(17)
 	})
 })
